Add new logic.
