+++ packages/commonwealth/server/routes/linkExistingAddressToChain.ts
@@ -8,6 +8,7 @@ import { ADDRESS_TOKEN_EXPIRES_IN } from '../config';
 import type { DB } from '../models';
 import { createRole, findOneRole } from '../util/roles';
 import { factory, formatFilename } from 'common-common/src/logging';
+import assertAddressOwnership from '../util/assertAddressOwnership';
 
 const log = factory.getLogger(formatFilename(__filename));
 
@@ -153,6 +154,9 @@ const linkExistingAddressToChain = async (
 
       addressId = newObj.id;
     }
+
+    // assertion check
+    await assertAddressOwnership(models, encodedAddress);
 
     const existingProfile = await models.OffchainProfile.findOne({
       where: { address_id: addressId },
